pyproject.toml
[tool.poetry]
# Project metadata
name = "synthpriv"  # The name of your project/package
version = "0.1.0"  # Project version
# Short description of the project
# (shown on PyPI and in package metadata)
description = "Differentially private synthetic data generator for tabular data."
# List of project authors
authors = ["Your Name <you@example.com>"]
# License identifier (e.g., MIT, Apache-2.0)
license = "MIT"

[tool.poetry.dependencies]
# Python version required for this project
python = ">=3.13,<4.0"
# Core dependencies
pandas = "^2.0.0"  # Data manipulation and analysis
ctgan = "^0.7.1"  # (Not compatible with 3.13, left for reference)
opacus = "^1.4.0"  # Differential privacy for PyTorch (not used with SDV)
matplotlib = "^3.8.0"  # Plotting library
streamlit = "^1.34.0"  # Web UI framework
typer = {extras = ["all"], version = "^0.9.0"}  # CLI framework
psycopg2-binary = "^2.9.9"  # PostgreSQL driver
python-dotenv = "^1.0.1"  # .env file support

[tool.poetry.group.dev.dependencies]
# Development dependencies (not required for production)
pytest = "^8.0.0"  # Testing framework
black = "^24.0.0"  # Code formatter
ruff = "^0.4.0"  # Linter
commitizen = "^3.20.0"  # Conventional commit tool

[build-system]
# Build system requirements
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
